requirements.txt
mysql-connector
flask
redis